% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_F_to_C}
\alias{convert_F_to_C}
\title{Convert Fahrenheit to Celsius}
\usage{
convert_F_to_C(x)
}
\arguments{
\item{x}{Numeric vector of temperatures in degrees Fahrenheit.}
}
\value{
A numeric vector of temperatures converted to degrees Celsius.
}
\description{
Converts a numeric vector of temperatures from degrees Fahrenheit (\u00B0F)
to degrees Celsius (\u00B0C) using the standard linear transformation.
}
\details{
This is an internal helper used when parsing HOBO logger files that
report temperature in \u00B0F.
}
\examples{
convert_F_to_C(32)   # 0
convert_F_to_C(212)  # 100

}
\keyword{internal}
