% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape regex special characters}
\usage{
escape_regex(x)
}
\arguments{
\item{x}{A character string to escape.}
}
\value{
A character string with regex special characters escaped.
}
\description{
Escape regex special characters
}
\keyword{internal}
